Sound/sst: port to 64bits kernel
BZ: 148118

Port intel sst to a 64 bits kernel.
The physical addresses being 64bits wide (at most), occurrences of
(unsigned int) for addresses should be replaced by phys_addr_t.

Moreover, a scrutiny should be performed to clearly separate
bus physical address space from LPE IP address space, and that is
not done by this patch.

As a first stage, it should be assumed that :
- audio fabric will still be 32 bits only
- as a consequence, it will only see the first 4GB of RAM

Furthermore, in a second stage, should be done if necessary :
- LPE IP enhancement to talk in 64bits wide on OCP bus
- LPE API should take 64bits for physical addresses
- Audio driver should handle all this gracefully

The goal of this patch is to make first stage possible.

+++ sound/soc/intel/sst/sst_drv_interface.c
@@ -767,7 +767,7 @@ static inline int sst_calc_tstamp(struct pcm_stream_info *info,
 {
 	size_t delay_bytes, delay_frames;
 	size_t buffer_sz;
-	size_t pointer_bytes, pointer_samples;
+	u32 pointer_bytes, pointer_samples;
 
 	pr_debug("mrfld ring_buffer_counter %llu in bytes\n",
 			fw_tstamp->ring_buffer_counter);
